docs: write doc comment for config.go
