--- internal/models/broker.go
@@ -1,8 +1,0 @@
-package models
-
-type Broker struct {
-	Username string `yaml:"user"`
-	Password string `yaml:"pass"`
-	Port     string `yaml:"port"`
-	Host     string `yaml:"host"`
-}
