--- internal/models/server.go
@@ -1,7 +1,0 @@
-// AnhCao 2024
-package models
-
-type Server struct {
-	Port string `yaml:"port"`
-	Host string `yaml:"host"`
-}
